# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat Serve API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/photos": {
      "get": {
        "summary": "List photos",
        "responses": {
          "200": {
            "description": "Photos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Photo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Get signed upload URL",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contentType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed URL"
          }
        }
      },
      "delete": {
        "summary": "Delete photo",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Photo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}